Split media.py into {track,video,album,playlist}.py
